feat(facets): remove filters on new query (#717)
EX-6196
