+++ packages/x-components/src/x-modules/facets/store/emitters.ts
@@ -1,5 +1,6 @@
 import { createStoreEmitters } from '../../../store/utils/store-emitters.utils';
 import { areFiltersDifferent } from '../../../utils/filters';
+import { isNewQuery } from '../../../utils/is-new-query';
 import { facetsXStoreModule } from './module';
 
 /**
@@ -11,5 +12,9 @@ export const facetsEmitters = createStoreEmitters(facetsXStoreModule, {
   SelectedFiltersChanged: {
     selector: (_, getters) => getters.selectedFilters,
     filter: areFiltersDifferent
+  },
+  FacetsQueryChanged: {
+    selector: state => state.query,
+    filter: isNewQuery
   }
 });
