+++ packages/x-components/src/x-modules/facets/store/module.ts
@@ -12,6 +12,7 @@ import { FacetGroupEntry, FacetsXStoreModule } from './types';
  */
 export const facetsXStoreModule: FacetsXStoreModule = {
   state: () => ({
+    query: '',
     filters: {},
     groups: {},
     facets: {},
@@ -47,6 +48,9 @@ export const facetsXStoreModule: FacetsXStoreModule = {
     },
     setFacet(state, facet: Facet) {
       Vue.set(state.facets, facet.id, facet);
+    },
+    setQuery(state, query) {
+      state.query = query;
     }
   },
   actions: {}
