+++ packages/x-components/src/x-modules/facets/store/types.ts
@@ -7,6 +7,8 @@ import { XActionContext, XStoreModule } from '../../../store';
  * @public
  */
 export interface FacetsState {
+  /** The current query {@link FacetsState.query}. */
+  query: string;
   /** Record of all available filters indexed by its id. */
   filters: Record<Filter['id'], Filter>;
   /** Record specifying the group each facet belongs to. */
@@ -81,6 +83,12 @@ export interface FacetsMutations {
    * @param filters - The filters to add.
    */
   setPreselectedFilters(filters: RawFilter[]): void;
+  /**
+   * Sets the {@link FacetsState.query} property.
+   *
+   * @param query - The new {@link FacetsState.query}.
+   */
+  setQuery(query: string): void;
   /**
    * Removes the facet from the {@link FacetsState.facets | facets} record.
    *
