+++ packages/x-components/tests/e2e/common/common-steps.spec.ts
@@ -227,6 +227,10 @@ Then(
   }
 );
 
+When('{string} is added to the search', (secondQuery: string) => {
+  cy.typeQuery(` ${secondQuery}`);
+});
+
 // Sort
 When('sort option {string} is selected from the sort dropdown', (sortOption: string) => {
   cy.getByDataTest(`sort-dropdown-toggle`).click();
