+++ packages/x-components/tests/e2e/url/url.spec.ts
@@ -14,3 +14,12 @@ When('selecting store {string}', (store: string) => {
   cy.getByDataTest('store-selector').getByDataTest('dropdown-toggle').click();
   cy.getByDataTest('store-selector').contains(store).click();
 });
+
+// Scenario 5
+And('url not contains parameter {string}', (key: string) => {
+  cy.location('search').should('not.contain', key);
+});
+
+When('{string} replaces current query', (query: string) => {
+  cy.replaceQuery(query);
+});
